dev: refactored config interface into separate file
